Refactor ChatAssistant; encapsulate workflow into ChatAgent and subclasses for models that necessitate special handling or separate workflows
